Improve messages and download error handling.
Download errors were sometimes not noticed by the main program, so it would proceed to verifying the download although it was incomplete and give inappropriate messages. I have rearranged things so it checks zsync_status itself.
Also improvements to many of the progress and error messages given, based on feedback.