Do not request data past end of target file.
The last block will be an incomplete block unless the file size is an
exact multiple of the blocksize. Previously zsync would ask for a full
block at the end of the file; most webservers would clip this to the
file size anyway, but we need not rely on that.
Reported by Aiko Barz https://bugs.debian.org/734185